Further restructuring, handling loading of assemblies in separate base class
