+++ .ci/unit-tests/Serialiser_Engine_Tests/AllTypesIObject.cs
@@ -8,15 +8,11 @@ using System.Threading.Tasks;
 
 namespace BH.Tests.Engine.Serialiser
 {
-    public static class AllTypesIObject
+    public class AllTypesIObject : BaseLoader
     {
         [Test]
         public static void CheckAllTypes()
         {
-            BH.Engine.Base.Compute.LoadAllAssemblies();
-            string regexFilter = $"^Revit_.*_20\\d\\d$";
-            BH.Engine.Base.Compute.LoadAllAssemblies("", regexFilter);
-
             List<Type> allTypes = BH.Engine.Base.Query.AllTypeList().Where(x => x.Namespace.StartsWith("BH.oM") || x.Namespace.StartsWith("BH.Revit.oM")).Where(x => !x.IsEnum).ToList();
 
             List<Type> nonIObjects = allTypes.Where(x => !typeof(IObject).IsAssignableFrom(x)).ToList();
